save investment into database
