+++ src/app/create_investment/[id]/page.tsx
