+++ src/components/investment_form.tsx
@@ -27,6 +27,7 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { type business } from "~/server/db/schema";
 import { isCreditCard, isDate } from "validator";
+import { getInvestmentSchema } from "~/app/create_investment/schema";
 
 function DialogCountdown({ isFormValid }: { isFormValid: boolean }) {
   const [countdown, setCountdown] = useState(3);
@@ -87,63 +88,7 @@ export function InvestingForm({
     businessData.businessID,
   );
 
-  const FormSchema = z.object({
-    terms: z.boolean().default(false),
-
-    cardNumber: z
-      .string()
-      .min(1, { message: "Credit card number is required." })
-      .refine((value) => isCreditCard(value.replace(/\s/g, "")), {
-        message: "Please enter a valid credit card number.",
-      })
-      .refine(
-        (value) => {
-          const digits = value.replace(/\s/g, "");
-          return digits.length >= 13 && digits.length <= 16;
-        },
-        {
-          message: "Credit card number must be between 13 and 16 digits.",
-        },
-      ),
-
-    expirationDate: z
-      .string()
-      .min(1, { message: "Expiration date is required." })
-      .refine(
-        (value) => {
-          const [month, year] = value.split("/");
-          if (!month || !year || month.length !== 2 || year.length !== 2)
-            return false;
-
-          const currentDate = new Date();
-          const expDate = new Date(2000 + parseInt(year), parseInt(month) - 1);
-
-          return !isNaN(expDate.getTime()) && expDate > currentDate;
-        },
-        {
-          message: "Please enter a valid expiration date (MM/YY)",
-        },
-      ),
-
-    cvv: z
-      .string()
-      .min(1, { message: "CVV is required." })
-      .refine((value) => /^\d{3,4}$/.test(value), {
-        message: "Please enter a valid CVV (3-4 digits)",
-      }),
-
-    amount: z
-      .number({
-        required_error: "Amount is required.",
-        invalid_type_error: "Amount must be a number.",
-      })
-      .min(businessData.min_investment!, {
-        message: `Minimum investment amount is ${businessData.min_investment} $`,
-      })
-      .positive({
-        message: "Amount must be positive.",
-      }),
-  });
+  const FormSchema = getInvestmentSchema(businessData.min_investment!);
 
   const form = useForm<z.infer<typeof FormSchema>>({
     resolver: zodResolver(FormSchema),
@@ -152,6 +97,7 @@ export function InvestingForm({
       cardNumber: "",
       expirationDate: "",
       cvv: "",
+      amount: businessData.min_investment!,
     },
     mode: "onBlur",
   });
